Renaming vars and changing ability usage args

+++ Assets/Scripts/CharacterStats.cs
@@ -6,7 +6,7 @@
     public readonly DamageType damageType;
     public readonly int damageIterations;
     public readonly int moveSpeed;
-    public readonly int initiative;
+    public readonly float initiative;
     public readonly int range;
 
     public CharacterStats(int maxHealth, int armor, int damage, int moveSpeed, int initiative, int range = 1, int damageIterations = 1, DamageType damageType = DamageType.normal)
